requirements.txt
Django>=2.0.2
setuptools
celery>=3.1.18
redis>=2.10.3
django-hijack
django-hijack-admin
django-extensions
django-debug-toolbar
django-bootstrap4
django-view-breadcrumbs
django-celery
django-celery-results
selenium
bs4
jinja2
wget
unidecode
django-import-export
requests
ntlm-auth
requests_ntlm
mysqlclient